Move Student and StudentCourse classes to data package
