+++ src/main/java/raisetech/StudentManagement5/Application.java
@@ -6,6 +6,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement5.data.Student;
+import raisetech.StudentManagement5.data.StudentCourse;
 
 @SpringBootApplication
 @RestController
