+++ src/main/java/raisetech/StudentManagement5/data/Student.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement5;
+package raisetech.StudentManagement5.data;
 
 import lombok.Getter;
