+++ src/main/java/raisetech/StudentManagement5/data/StudentCourse.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement5;
+package raisetech.StudentManagement5.data;
 
 import java.time.LocalDateTime;
 import lombok.Getter;
